sec9-gist-of-RAG/pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=0.3.31",
    "langchain-pinecone>=0.2.11",
    "langchainhub>=0.1.21",
    "ollama>=0.6.1",
    "python-dotenv>=1.1.1",
]
